# switch to using json sanitizer for admin counters, gauges, and text-readouts