--- a/source/server/admin/BUILD
+++ b/source/server/admin/BUILD
@@ -117,6 +117,7 @@ envoy_cc_library(
         ":utils_lib",
         "//source/common/buffer:buffer_lib",
         "//source/common/html:utility_lib",
+        "//source/common/json:json_sanitizer_lib",
         "//source/common/stats:histogram_lib",
     ],
 )