--- a/source/server/admin/stats_render.cc
+++ b/source/server/admin/stats_render.cc
@@ -4,11 +4,6 @@
 #include "source/common/json/json_sanitizer.h"
 #include "source/common/stats/histogram_impl.h"
 
-namespace {
-// This value found by iterating in stats_handler_speed_test.cc, maximizing the
-// performance of the Json tests, and then rounding to a power of 2.
-constexpr uint64_t JsonStatsFlushCount = 64;
-} // namespace
 namespace Envoy {
 
 using ProtoMap = Protobuf::Map<std::string, ProtobufWkt::Value>;