--- a/source/server/admin/stats_render.cc
+++ b/source/server/admin/stats_render.cc
@@ -4,6 +4,14 @@
 #include "source/common/json/json_sanitizer.h"
 #include "source/common/stats/histogram_impl.h"
 
+namespace {
+constexpr absl::string_view JsonNameTag = "{\"name\":\"";
+constexpr absl::string_view JsonValueTag = "\",\"value\":";
+constexpr absl::string_view JsonValueTagQuote = "\",\"value\":\"";
+constexpr absl::string_view JsonCloseBrace = "}";
+constexpr absl::string_view JsonQuoteCloseBrace = "\"}";
+} // namespace
+
 namespace Envoy {
 
 using ProtoMap = Protobuf::Map<std::string, ProtobufWkt::Value>;
@@ -90,16 +98,18 @@ StatsJsonRender::StatsJsonRender(Http::ResponseHeaderMap& response_headers,
 // buffer once we exceed JsonStatsFlushCount stats.
 void StatsJsonRender::generate(Buffer::Instance& response, const std::string& name,
                                uint64_t value) {
-  addJson(response, absl::StrCat("{\"name\":\"", Json::sanitize(name_buffer_, name),
-                                 "\",\"value\":", value, "}"));
+  response.addFragments({delim_, JsonNameTag, Json::sanitize(name_buffer_, name),
+      JsonValueTag, std::to_string(value), JsonCloseBrace});
+  delim_ = ",";
 }
 
 // Buffers a JSON fragment for a text-readout stat, flushing to the response
 // buffer once we exceed JsonStatsFlushCount stats.
 void StatsJsonRender::generate(Buffer::Instance& response, const std::string& name,
                                const std::string& value) {
-  addJson(response, absl::StrCat("{\"name\":\"", Json::sanitize(name_buffer_, name),
-                                 "\",\"value\":\"", Json::sanitize(value_buffer_, value), "\"}"));
+  response.addFragments({delim_, JsonNameTag, Json::sanitize(name_buffer_, name),
+      JsonValueTagQuote, Json::sanitize(value_buffer_, value), JsonQuoteCloseBrace});
+  delim_ = ",";
 }
 
 // In JSON we buffer all histograms and don't write them immediately, so we
@@ -140,9 +150,6 @@ void StatsJsonRender::generate(Buffer::Instance&, const std::string& name,
 // Since histograms are buffered (see above), the finalize() method generates
 // all of them.
 void StatsJsonRender::finalize(Buffer::Instance& response) {
-  if (!stats_array_.empty()) {
-    flushStats(response);
-  }
   if (histogram_array_->values_size() > 0) {
     ProtoMap& histograms_obj_container_fields = *histograms_obj_container_.mutable_fields();
     if (found_used_histogram_) {
@@ -163,41 +170,10 @@ void StatsJsonRender::finalize(Buffer::Instance& response) {
     // untrappable error message to stdout) if it receives an invalid input, so
     // we exclude that here.
     if (!str.empty()) {
-      addJson(response, str);
+      response.addFragments({delim_, str});
     }
   }
   response.add("]}");
-}
-
-// Flushes all stats that were buffered in addJson() above.
-void StatsJsonRender::flushStats(Buffer::Instance& response) {
-  ASSERT(!stats_array_.empty());
-  // TODO(jmarantz): when #20428 lands we can serialize the scalar stat
-  // values without using the slow protobuf serializer.
-  const std::string json_array = MessageUtil::getJsonStringFromMessageOrDie(
-      ValueUtil::listValue(stats_array_), false /* pretty */, true);
-  stats_array_.clear();
-
-  // We are going to wind up with multiple flushes which have to serialize as
-  // a single array, rather than a concatenation of multiple arrays, so we add
-  // those in the constructor and finalize() method, strip off the "[" and "]"
-  // from each buffered serialization.
-  ASSERT(json_array.size() >= 3);
-  ASSERT(json_array[0] == '[');
-  ASSERT(json_array[json_array.size() - 1] == ']');
-  addJson(response, absl::string_view(json_array).substr(1, json_array.size() - 2));
-}
-
-// Adds a json fragment of scalar stats to the response buffer, including a
-// "," delimiter if this is not the first fragment.
-void StatsJsonRender::addJson(Buffer::Instance& response, absl::string_view json) {
-  ASSERT(!json.empty());
-  if (first_) {
-    response.add(json);
-    first_ = false;
-  } else {
-    response.addFragments({",", json});
-  }
 }
 
 // Summarizes the buckets in the specified histogram, collecting JSON objects.